+++ frontend/main-app/src/components/SubmissionViewer/index.ts
@@ -1,0 +1,1 @@
+export { SubmissionViewer } from "./SubmissionViewer";
